requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
protobuf==3.19.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-jwt==3.3.2
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
